Updated to MC1.12.2-1.3.1
This version is the start of the new original RTP system! This system will be safer, faster and more optimized than the old one ( /spreadplayers command ). Remember that this new system it's still on beta so if you find any issue send me a message on discord, to change from the old system to the new one just change the option use-original to true. Also this version was not possible to make without the incredible code of GeiloUtils (https://github.com/Geilokowski/GeiloUtils). Here is a detailed change-log:
- Created a new system based on GeiloUtils of the /rtp command
- Added the ability to choose between the New System and Old System in the config (use-original)
- Solved minor issues

How to update:
1. Delete the old .jar file from the mods folder of the server root
2. Move the new .jar to the folder mods of the server root